+++ controllers/comment.js
@@ -25,21 +25,30 @@ exports.getComment = async (req, res, next) => {
     } catch(err) { next(err) }
 }
 
-exports.addComment = async (req, res, next) => {
+exports.addComment = async (user_Id, req, res, next,) => {
     
     try {
-        const {body, rate} = req.body
-        if(!body || !rate) {
+        const {body, rate, chocolate_Id} = req.body
+        if(!body || !rate || !chocolate_Id) {
             throw new RequestError("Missing Data")
         }
+
+        const chocolate = await DB.Chocolate.findOne({ where: { id: chocolate_Id }, raw: true })
+        if(chocolate === null) {
+            throw new CommentError(`the chocolate of the comment don't exists`)
+        }
         
-        const comment = await Comment.create(req.body)
+        const commentWithUser = {
+            body, rate, chocolate_Id, 
+            user_comment_Id: user_Id
+        }
+        const comment = await Comment.create(commentWithUser)
         return res.json({ message: "Comment Created", data: comment })
         
     } catch(err) { next(err) }
 }
 
-exports.updateComment = async (req, res, next) => {
+exports.updateComment = async (user_Id, req, res, next) => {
     try {
         const commentId = parseInt(req.params.id)
         if(!commentId) {
@@ -53,20 +62,29 @@ exports.updateComment = async (req, res, next) => {
             throw new CommentError("This comment does not exist !")
         }
 
+        if (comment.user_comment_Id !== user_Id) {
+            return res.json(404).json({ message: "You don't have the right for this"})
+        }
+
         await Comment.update(req.body, { where: { id: commentId } })
         return res.json({ message: "Comment Updated" })
 
     } catch(err) { next(err) }
 }
 
-exports.deleteComment = async (req, res, next) => {
+exports.deleteComment = async (user_Id, req, res, next) => {
     try {
         let commentId = parseInt(req.params.id)
         if(!commentId) {
             // return res.json(400).json({ message: "Missing Parameter"})
             throw new RequestError("Missing Data")
         }
-        
+
+        const comment = await Comment.findOne({ where: {id: commentId}, raw: true})
+        if (comment.user_comment_Id !== user_Id) {
+            return res.json(404).json({ message: "You don't have the right for this"})
+        }
+
         await Comment.destroy({ where: {id: commentId}, force: true})
         return res.status(204).json({})
