+++ controllers/user.js
@@ -31,7 +31,7 @@ exports.getUser = async (req, res, next) => {
     }
 }
 
-exports.addUser = async (req, res, next) => {
+exports.addUser = async (user_Id, req, res, next) => {
     try {
         const {speudo, email, password } = req.body
         if(!speudo || !email || !password) {
@@ -47,9 +47,13 @@ exports.addUser = async (req, res, next) => {
         }
 
         const hash = await bcrypt.hash(password, parseInt(process.env.BCRYPT_SALT_ROUND))
-        req.body.password = hash
+        // req.body.password = hash
 
-        const userCrypted = await User.create(req.body)
+        let userCrypted =  {
+            speudo, email, user_Id,
+            password: hash 
+        }
+        userCrypted = await User.create(req.body)
         return res.json({message: "UserCreated", data: userCrypted})
         
     } catch(err) {
@@ -58,7 +62,7 @@ exports.addUser = async (req, res, next) => {
     }
 }
 
-exports.updateUser = async (req, res, next) => {
+exports.updateUser = async (user_Id, req, res, next) => {
     try {
         let userId = parseInt(req.params.id)
         if(!userId) {
@@ -72,7 +76,11 @@ exports.updateUser = async (req, res, next) => {
             // return res.status(404).json({ message: "This user does not exist !"})
             throw new UserError("This user does not exist !",0)
         }
-        
+
+        if (user.id !== user_Id) {
+            return res.json(404).json({ message: "You don't have the right for this"})
+        }
+
         await User.update(req.body, { where: { id: userId } })
         return res.json({ message: "User Updated" })
 
@@ -82,13 +90,18 @@ exports.updateUser = async (req, res, next) => {
     }
 }
 
-exports.untrashUser = async (req, res, next) => {
+exports.untrashUser = async (user_Id, req, res, next) => {
     try {
         let userId = parseInt(req.params.id)
         if(!userId) {
             throw new RequestError("Missing parameter")
         }
-    
+        
+        const user = await User.findOne({ where: {id: userId}, raw: true})
+        if (user.id !== user_Id) {
+            return res.json(404).json({ message: "You don't have the right for this"})
+        }
+
         await User.restore({ where : {id: userId}})
         return res.status(204).json({})
     } catch(err) {
@@ -96,11 +109,16 @@ exports.untrashUser = async (req, res, next) => {
     }
 }
 
-exports.trashUser = async (req, res, next) => {
+exports.trashUser = async (user_Id, req, res, next) => {
     try {
         let userId = parseInt(req.params.id)
         if(!userId) {
            throw new RequestError("Missing parameter")
+        }
+
+        const user = await User.findOne({ where: {id: userId}, raw: true})
+        if (user.id !== user_Id) {
+            return res.json(404).json({ message: "You don't have the right for this"})
         }
         
         await User.destroy({ where: {id: userId}})
@@ -110,13 +128,18 @@ exports.trashUser = async (req, res, next) => {
     }
 }
 
-exports.deleteUser = async (req, res, next) => {
+exports.deleteUser = async (user_Id, req, res, next) => {
     try {
         let userId = parseInt(req.params.id)
         if(!userId) {
             throw new RequestError("Missing parameter")
         }
         
+        const user = await User.findOne({ where: {id: userId}, raw: true})
+        if (user.id !== user_Id) {
+            return res.json(404).json({ message: "You don't have the right for this"})
+        }
+
         await User.destroy({ where: {id: userId}, force: true})
         return res.status(204).json({})
